+++ reservation-api/src/main/java/ca/ulaval/glo4002/reservation/domain/fullcourse/recipe/allergy/AllergyDessert.java
@@ -1,11 +1,13 @@
 package ca.ulaval.glo4002.reservation.domain.fullcourse.recipe.allergy;
 
+import java.math.BigDecimal;
+
 import ca.ulaval.glo4002.reservation.domain.fullcourse.Ingredient;
 import ca.ulaval.glo4002.reservation.domain.fullcourse.IngredientName;
 import ca.ulaval.glo4002.reservation.domain.fullcourse.Recipe;
 
 public class AllergyDessert extends Recipe {
-  private static final double BACON_QUANTITY = 5;
+  private static final BigDecimal BACON_QUANTITY = BigDecimal.valueOf(5);
 
   public AllergyDessert() {
     ingredients.add(new Ingredient(IngredientName.BACON, BACON_QUANTITY));
