+++ reservation-api/src/main/java/ca/ulaval/glo4002/reservation/infra/inmemory/IngredientQuantityRepository.java
@@ -1,5 +1,6 @@
 package ca.ulaval.glo4002.reservation.infra.inmemory;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.Map;
@@ -12,7 +13,7 @@ import ca.ulaval.glo4002.reservation.domain.util.MapUtil;
 
 public class IngredientQuantityRepository {
 
-  private final Map<LocalDate, Map<IngredientName, Double>> ingredientsQuantityPerDay = new HashMap<>();
+  private final Map<LocalDate, Map<IngredientName, BigDecimal>> ingredientsQuantityPerDay = new HashMap<>();
 
   private final ReservationIngredientCalculator reservationIngredientCalculator;
 
@@ -25,16 +26,16 @@ public class IngredientQuantityRepository {
   }
 
   public void updateIngredientsQuantity(Reservation reservation) {
-    Map<IngredientName, Double> currentIngredientsQuantity = getIngredientsQuantity(LocalDate.from(reservation.getDinnerDate()));
+    Map<IngredientName, BigDecimal> currentIngredientsQuantity = getIngredientsQuantity(LocalDate.from(reservation.getDinnerDate()));
 
-    Map<IngredientName, Double> reservationIngredientsQuantity = reservationIngredientCalculator.getReservationIngredientsQuantity(reservation);
-    Map<IngredientName, Double> updatedIngredientsQuantity = MapUtil.merge(reservationIngredientsQuantity,
-                                                                           currentIngredientsQuantity);
+    Map<IngredientName, BigDecimal> reservationIngredientsQuantity = reservationIngredientCalculator.getReservationIngredientsQuantity(reservation);
+    Map<IngredientName, BigDecimal> updatedIngredientsQuantity = MapUtil.merge(reservationIngredientsQuantity,
+                                                                               currentIngredientsQuantity);
     ingredientsQuantityPerDay.put(LocalDate.from(reservation.getDinnerDate()),
                                   updatedIngredientsQuantity);
   }
 
-  public Map<IngredientName, Double> getIngredientsQuantity(LocalDate date) {
+  public Map<IngredientName, BigDecimal> getIngredientsQuantity(LocalDate date) {
     if (ingredientsQuantityPerDay.containsKey(date)) {
       return ingredientsQuantityPerDay.get(date);
     }
@@ -42,8 +43,8 @@ public class IngredientQuantityRepository {
     return ingredientsQuantityPerDay.get(date);
   }
 
-  public Map<LocalDate, Map<IngredientName, Double>> getIngredientsQuantity(ReportPeriod reportPeriod) {
-    Map<LocalDate, Map<IngredientName, Double>> ingredientsQuantity = new HashMap<>();
+  public Map<LocalDate, Map<IngredientName, BigDecimal>> getIngredientsQuantity(ReportPeriod reportPeriod) {
+    Map<LocalDate, Map<IngredientName, BigDecimal>> ingredientsQuantity = new HashMap<>();
     for (LocalDate date : reportPeriod.getAllDaysOfPeriod()) {
       if (ingredientsQuantityPerDay.containsKey(date)) {
         ingredientsQuantity.put(date, ingredientsQuantityPerDay.get(date));
