+++ reservation-api/src/test/java/ca/ulaval/glo4002/reservation/api/report/assembler/IngredientReportInformationDtoAssemblerTest.java
@@ -22,8 +22,8 @@ import ca.ulaval.glo4002.reservation.domain.report.unit.UnitReportDay;
 public class IngredientReportInformationDtoAssemblerTest {
   private static final IngredientName BUTTERNUT_SQUASH_NAME = IngredientName.BUTTERNUT_SQUASH;
   private static final IngredientName CHOCOLATE_NAME = IngredientName.CHOCOLATE;
-  private static final double A_QUANTITY = 2.0;
-  private static final double ANOTHER_QUANTITY = 3.0;
+  private static final BigDecimal A_QUANTITY = BigDecimal.valueOf(2.0);
+  private static final BigDecimal ANOTHER_QUANTITY = BigDecimal.valueOf(3.0);
   private static final BigDecimal A_TOTAL_PRICE = BigDecimal.TEN;
   private static final BigDecimal ANOTHER_TOTAL_PRICE = BigDecimal.ONE;
